% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jatimuser.R
\name{get_dataset}
\alias{get_dataset}
\title{Mengambil Dataset Berdasarkan Nama}
\usage{
get_dataset(dataset_name)
}
\arguments{
\item{dataset_name}{Character. Nama sheet (dataset) yang ingin diambil.}
}
\value{
Data frame.
}
\description{
Mengambil dataset dari Google Spreadsheet berdasarkan
nama sheet.
}
