% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jatimuser.R
\docType{package}
\name{jatimuseR-package}
\alias{jatimuseR}
\alias{jatimuseR-package}
\title{jatimuseR: Akses Dataset Badan Pusat Statistik Provinsi Jawa Timur Siap Pakai}
\description{
Kumpulan fungsi untuk mengambil dataset dari Google Spreadsheet
secara read-only (tanpa autentikasi) untuk keperluan analisis data
Badan Pusat Statistik Provinsi Jawa Timur.
}
\author{
\strong{Maintainer}: Joko Nursiyono \email{email@anda.com}

}
\keyword{internal}
