% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jatimuser.R
\name{list_datasets}
\alias{list_datasets}
\title{Menampilkan Daftar Dataset (Nama Sheet)}
\usage{
list_datasets()
}
\value{
Character vector berisi nama dataset.
}
\description{
Mengembalikan daftar nama sheet yang tersedia
pada Google Spreadsheet jatimuseR.
}
